refactor structure and refine configs
