+++ examples/test_data_gen/config.ini.example
@@ -1,0 +1,34 @@
+; config.ini
+; This is a sample configuration file
+
+[DEFAULT]
+; The DEFAULT section provides default values for all other sections.
+; Configure both 'document_folder' and 'document_chunk_size' if you require document splitting.
+; However, if you wish to bypass the document split process, simply provide the 'document_node_file', which is a JSONL file.
+; When all these parameters are configured, the system will primarily use the 'document_node_file'
+document_folder = "<your-document-folder-abspath>"
+document_chunk_size = 1024
+document_node_file = "<your-node-file-abspath>"
+
+; Test data gen flow configs
+flow_folder = "<your-test-data-gen-flow-folder-abspath>" ; There is must one flow.dag.yaml file under this folder as entry
+connection_name = "<your-connection-name>"
+
+
+[LOCAL]
+; This section is for local test data generation related configuration.
+output_folder = "<your-output-folder-abspath>"
+flow_batch_run_size = 10
+
+
+[CLOUD]
+; This section is for cloud test data generation related configuration.
+subscription_id = "<your-sub-id>"
+resource_group = "<your-resource-group>"
+workspace_name = "<your-workspace-name>"
+aml_cluster = "<your-compute-name>"
+
+; Parallel run step configs
+prs_instance_count = 2
+prs_mini_batch_size = "10kb"
+prs_max_concurrency_per_instance = 10
