--- examples/test_data_gen/src/test_data_gen_local/config.ini
@@ -1,11 +1,0 @@
-; should_skip_doc_split = True
-documents_folder = "D:\proj\github\ms\promptflow\docs\how-to-guides"
-document_chunk_size = 1024
-; cspell: ignore luyao
-document_nodes_output_path = "D:\proj\github\ms\promptflow\examples\test_data_gen\document-nodes-test.jsonl"
-; test data gen flow configs
-flow_folder = "D:\proj\github\ms\promptflow\examples\test_data_gen\construct_test_data_flow"
-flow_batch_run_size = 10
-connection_name = "azure_open_ai_connection"
-; test data output path
-test_data_output_path = "D:\proj\github\ms\promptflow\examples\test_data_gen\test_data"
