+++ examples/test_data_gen/test_data_gen/construct_test_data_flow/generate_ground_truth_prompt.jinja2
